requirements.txt
# Bharat Law — RAG Legal Chatbot

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# PDF parsing
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.3.0

# Embeddings & reranking
sentence-transformers>=3.0.0
torch>=2.3.0
numpy<=1.24.3

# Vector DB
pinecone>=4.0.0

# LLM
google-generativeai>=0.7.0

# Utilities
httpx>=0.27.0
slowapi>=0.1.9
aiofiles>=23.2.1
tenacity>=8.3.0   # retry logic

# Observability / storage
# SQLite is stdlib — no extra install needed

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0     # also used for TestClient
streamlit>=1.35.0
